Consolidate with pickUtils
