+++ src/commands/image/imageSource/containerRegistry/acr/AcrRepositoriesListStep.ts
@@ -7,9 +7,10 @@ import { KnownActiveRevisionsMode } from "@azure/arm-appcontainers";
 import { uiUtils } from "@microsoft/vscode-azext-azureutils";
 import { nonNullValue } from "@microsoft/vscode-azext-utils";
 import { type QuickPickItem } from "vscode";
-import { acrDomain, currentlyDeployedPickDescription, noMatchingResourcesQp } from "../../../../../constants";
+import { acrDomain, noMatchingResourcesQp } from "../../../../../constants";
 import { createContainerRegistryClient } from "../../../../../utils/azureClients";
 import { parseImageName } from "../../../../../utils/imageNameUtils";
+import { currentlyDeployedPickDescription } from "../../../../../utils/pickUtils";
 import { type ContainerRegistryImageSourceContext } from "../ContainerRegistryImageSourceContext";
 import { RegistryRepositoriesListStepBase } from "../RegistryRepositoriesListBaseStep";
 import { getLatestContainerAppImage } from "../getLatestContainerImage";
